Filter slope and mean data to remove scarce data ranges before fitting polynomial

Filter slope and mean data to remove scarce data ranges before fitting polynomial

Usage

filter_slopes_and_mean(
  data,
  window_size,
  min_obs,
  max_filter = TRUE,
  min_filter = TRUE
)

Arguments

data

output from longpoly::get_slopes_and_mean() (or any tibble with columns "id", "performance_slope", & "performance_mean")

window_size

the difference added to /subtracted from the minuimum/maximum values to subset the data. The number of observations in this subset is counted

min_obs

the minimum number of observations required in the subset defined by window_size. If the fewer observations exist, data in these ranges are filtered out

max_filter

filter highest scores? default = TRUE

min_filter

filter lowest scores? default = TRUE

Value

a dataset containing slope and mean values with sparse data removed (if requested and identified)

Examples

longpoly_example_data %>% nrow
#> [1] 2000

filtered_example_data <- filter_slopes_and_mean(
data = longpoly_example_data,
window_size = 0.25,
min_obs = 20,
max_filter = TRUE,
min_filter = TRUE
)
#> [1] "number of observations in max window:  4"
#> [1] "user defined miniumum number of observations:  20"
#> [1] "data in max window removed"
#> [1] "number of observations in min window:  4"
#> [1] "user defined miniumum number of observations:  20"
#> [1] "data in min window removed"

filtered_example_data %>% nrow
#> [1] 1992