A data-driven approach to choosing order of polynomial model describing the relationship between linear rate of change and mean score in longitudinal data

A data-driven approach to choosing order of polynomial model describing the relationship between linear rate of change and mean score in longitudinal data

Usage

find_polynomial(data, test_proportion = 1/3, idcol = "id", max_order = 6, x)

Arguments

data

output from longpoly::get_slopes_and_mean() (or any tibble with columns "id", "performance_slope", & "performance_mean")

test_proportion

the proportion of participants to allocate to the test set. default = 1/3

idcol

the column name corresponding to the ID variable in data. Default = "id"

max_order

the maximum order for polynomial models to be tested. Polynomial models for 1:max_order will be tested. default = 6

x

Polynomials are selected after finding the maximum proportion of variance explained (PVE) in test data across all models, and then finding the most parsimonious model with a PVE that is at least max(PVE) - x

Value

a list containing:

'polynomial_results' - a tibble with columns for the order of the model tested and the corresponding pve in the train and test data

'selected_order' - the order of the selected model (based on the max(PVE) - x criteria)

'selected_model_coefficients' - coefficients of the selected model (based on the max(PVE) - x criteria)

'train_ids' — a character vector of ids allocated to the train dataset in model development

'test_ids' — a character vector of ids allocated to the test dataset in model development

Examples

find_poly_results <- find_polynomial(data = longpoly_example_data, test_proportion = 1/3, max_order = 6, x = 0.05)

# View proportion of variance explained in the train and test data for each polynomial tested
find_poly_results$polynomial_results
#> # A tibble: 6 × 3
#>   order pve_in_train_data pve_in_test_data
#>   <int>             <dbl>            <dbl>
#> 1     1            0.0554           0.0591
#> 2     2            0.0762           0.0674
#> 3     3            0.158            0.228 
#> 4     4            0.159            0.238 
#> 5     5            0.194            0.210 
#> 6     6            0.203            0.327 

# View the order of the selected model
find_poly_results$selected_order
#> [1] 6

# View the coefficients of the selected model
find_poly_results$selected_model_coefficients
#>                            (Intercept) poly(performance_mean, 6, raw = TRUE)1 
#>                          -1.310842e-01                           1.929380e-01 
#> poly(performance_mean, 6, raw = TRUE)2 poly(performance_mean, 6, raw = TRUE)3 
#>                          -6.407160e-02                          -6.530676e-02 
#> poly(performance_mean, 6, raw = TRUE)4 poly(performance_mean, 6, raw = TRUE)5 
#>                           1.016220e-02                           5.302042e-03 
#> poly(performance_mean, 6, raw = TRUE)6 
#>                          -5.921981e-05 

# The IDs of the participants allocated to the train and test sets are stored in the following vectors
find_poly_results$train_ids |> head()
#> [1] 1197 1431  460   63  485 1894
find_poly_results$test_ids |> head()
#> [1]  1  2  3  9 10 11