Fit a polynomial of a specified order to describe the relationship between linear rate of change and mean score in longitudinal data

Fit a polynomial of a specified order to describe the relationship between linear rate of change and mean score in longitudinal data

Usage

implement_polynomial(
  data,
  performance_metric = "mean",
  order = 3,
  floor_effects = FALSE,
  floor_range = NULL
)

Arguments

data

output from longpoly::get_slopes_and_mean() (or any tibble with columns "id", "performance_slope", & "performance_mean" or "performance_bl")

performance_metric

which measure of performance is being used? Must be either "mean" (in which case performance_mean column must be in data) or "baseline" (data must contain performance_bl). Default = "mean"

order

the order of the polynomial model (recommended to be selected on the scree plot output from longpoly::test_polynomial()). default = 3

floor_effects

test for floor effects? These are identified by finding the mean value of the performance variable that is associated with the fastest rate of decline in the impaired range (specified by floor_range()). default = FALSE

floor_range

(required if floor_effects = TRUE). the 'impaired' range over which to search for the value of "performance_mean" that returns the minimum value for "performance_slope". expects a vector of length two with the first value providing the lower bound of the range and the second value the upper.

Value

a list containing:

model_formula — the formula of the final model

final_data — a tibble with columns for "id", "performance_slope", either "performance_mean" or "performance_bl" (depending on performance_metric), "predicted_slope", and "residual." If floor_effects = TRUE, an additional column "floor_effects" is appended with either keep/remove reflecting whether the record should be removed based on the identified performance_mean threshold for floor effects

threshold — if floor_effects = TRUE, this records the threshold used for floor effect classifications

Examples

poly_out <-
implement_polynomial(
  data = longpoly_example_data,
  performance_metric = "mean",
  order = 3,
  floor_effects = TRUE,
  floor_range = c(min(longpoly_example_data$performance_mean), 0)
)

# View the equation of the final model
poly_out$model_formula
#> [1] "y = -0.164 +0.129x -0.005x^2 -0.019x^3"

# View the final data
poly_out$final_data |> head()
#> # A tibble: 6 × 6
#>      id performance_mean performance_slope predicted_slope residual
#>   <int>            <dbl>             <dbl>           <dbl>    <dbl>
#> 1     1            0.235            0.127           -0.134   0.261 
#> 2     2           -0.331            0.0715          -0.207   0.278 
#> 3     3           -0.312           -0.215           -0.204  -0.0105
#> 4     4           -2.30             0.126           -0.255   0.381 
#> 5     5           -0.171           -0.215           -0.186  -0.0289
#> 6     6            0.140           -0.219           -0.146  -0.0725
#> # ℹ 1 more variable: floor_effects <fct>

# View the selected threshold
 poly_out$threshold
#> [1] -1.592723

# Other example uses:

# implement_polynomial(example_slope_mean_data,
#                      floor_effects = FALSE)
#
# # if floor_effects = TRUE,
# it may be sensible to use the minumum performance_mean value as the lower bound of the range,
# e.g.:
# example_slope_mean_data <-
#   longpoly::get_slopes_and_mean(
#     example_data,
#     id_col = "id",
#     outcome_col = "memory_test",
#     time_col = "time"
#   ))
#
#
# implement_polynomial(
#   example_slope_mean_data,
#   floor_effects = TRUE,
#   floor_range = c(min(
#     example_slope_mean_data$performance_mean
#   ), 0)
# )
#
#
# # if floor_effects = FALSE,
# it is straightforward to pipe in the output from longpoly::get_slopes_and_mean(),
# e.g.:
# get_slopes_and_mean(example_data,
#                    id_col = "id",
#                    outcome_col = "memory_test",
#                    time_col = "time") %>%
#   implement_polynomial(., floor_effects = FALSE)